re-org
getting the repo in more order
