+++ .gitignore
@@ -1,0 +1,14 @@
+
+previois version/view.html
+previois version/view.html
+previois version/view.html
+previois version/view.html
+previois version/nginx.config
+previois version/newmakeup.sql
+previois version/makeup.html
+previois version/index.html
+previois version/gowiki3.go
+previois version/gowiki3.exe
+previois version/edit.html
+previois version/Dockerfile
+previois version/dbwebpage.go
