+++ README.md
@@ -1,3 +1,41 @@
 # makeup-app
 
 This is an app for organizing makeup.  
+
+Makeup database app where you can track your usage
+I want customizable NFC tags for the cosmetics & a stand alone art object that has a readout display NFC reader, I wanted it to play a sound but if it has network capabilities that has the potential to be turned into a mic and idk if I want that so maybe just lights? 
+I want the display to show a fortune when you scan a product. 
+the app tracks what you put on when so if you wanna see if that 18 hr lip lasts 18 hr or not. 
+the app has like collections or orders of operations so like if you want to doa 42 setp sink care system you can like keep track of that. 
+you can have like wish lists and like follow other peoples collections and like private or public collections you have to have an account to like see the content; 
+would be a web app but accessable from a browser. 
+kind of like pastel and black probably. 
+
+tags, search function
+
+end to end encryption? 
+
+written in go
+casandra
+docker
+
+frontend
+react.js
+php?
+
+look into gitlab for ci/cd 
+
+probably serve on AWS
+
+
+this would have to be in python. 
+PN532 NFC HAT for Raspberry Pi, I2C / SPI / UART
+https://www.waveshare.com/wiki/PN532_NFC_HAT
+https://youtu.be/xoI7-X6V4qc
+
+
+
+photo guidelines:
+no nudity
+no offensive matererals
+neutral background for product
