+++ current_project/index.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html>
+	<head>
+		<link rel="stylesheet" href="http://cdn.jsdelivr.net/npm/semantic-ui@2.4.2/dist/semantic.min.css">
+		<title>Make up rating app</title>
+	</head>
+	<body>
+		<div id="app" />
+	</body>
+</html>
+
+<!-- I'm not sure the div app is correct, but I really need to look at this in a browser to see if working -->
