+++ current_project/styles.css
@@ -1,0 +1,85 @@
+body {
+  width: 100%;
+  height: 100vh;
+  margin: 0;
+  background-image: linear-gradient(0deg, #6f32a8, #db81cf, #6f32a8);
+  color: #2f2e30;
+  font-family: Impact;
+  font-size: 20px;
+}
+
+h1, p {
+  margin: 1em auto;
+  text-align: center;
+}
+
+form {
+  width: 60vw;
+  max-width: 500px;
+  min-width: 300px;
+  margin: 0 auto;
+  padding-bottom: 2em;
+}
+
+fieldset {
+  border: none;
+  padding: 2rem 0;
+  border-bottom: 3px solid #3b3b4f;
+}
+
+fieldset:last-of-type {
+  border-bottom: none;
+}
+
+label {
+  display: block;
+  margin: 0.5rem 0;
+}
+
+input,
+textarea,
+select {
+  margin: 10px 0 0 0;
+  width: 100%;
+  min-height: 2em;
+}
+
+input, textarea {
+  border-radius: 25px;
+  background-color: #3b3b4f;
+  border: 3px dotted #0a0a23;
+  color: #ffffff;
+}
+
+.inline {
+  width: unset;
+  margin: 0 0.5em 0 0;
+  vertical-align: middle;
+}
+
+input[type="submit"] {
+  display: block;
+  width: 60%;
+  margin: 1em auto;
+  height: 2em;
+  font-size: 1.1rem;
+  background-color: #3b3b4f;
+  border-color: white;
+  min-width: 300px;
+}
+
+input[type="file"] {
+  padding: 1px 2px;
+}
+
+a:link {
+  color: #3b3b4f;
+}
+
+a:visited {
+  color: #6f32a8;
+}
+
+a:hover {
+  color: #0a0a23;
+}
