+++ previous_version/Dockerfile
@@ -1,0 +1,4 @@
+FROM nginx
+
+COPY nginx.config /etc/nginx/conf.d/default.conf
+COPY index.html /usr/share/nginx/html
