+++ previous_version/edit.html
@@ -1,0 +1,11 @@
+<h1>Editing {{.Title}}</h1>
+
+<form action="/save/{{.Title}}" method="POST">
+	<div>
+		<textarea name="body" rows="20" cols="80">{{printf "%s" .Body}}
+		</textarea>
+	</div>
+	<div>
+		<input type="submit" value="Save">
+	</div>
+</form>
